[XLA:CPU] Fix crash due to OOM in XLA's custom convolution algorithm.
This commit adds a flag to set the maximum workspace size (i.e. the maximum amount of memory that temporary buffers can use).

At the moment this flag is used only in the custom transposed convolution algorithm. If the size of convolution matrix would exceed the maximum allowed workspace size, we fallback to the generic, slower algorithm.

PiperOrigin-RevId: 704221971